backendAPIs/fast-api/requirements.txt
alembic==1.11.1
black==24.10.0
fastapi==0.115.4
httpx==0.27.2
isort==5.13.2
psycopg2-binary==2.9.7
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
uvicorn==0.32.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.21.0